+++ examples/js/sort/demo.js
@@ -1,4 +1,5 @@
 /* eslint max-len: 0 */
+require('./style.css');
 import React from 'react';
 import SortTable from './sort-table';
 import MultiSortTable from './multi-sort-table';
@@ -13,6 +14,7 @@ import CustomCaretSortTable from './custom-caret-sort-table';
 import ExternalMultiSort from './manage-multi-sort-external-table';
 import DefaultASCSortTable from './default-asc-sort-table';
 import CleanSortedTable from './clean-sorted-table';
+import SortStyleTable from './sort-style-table';
 
 class Demo extends React.Component {
   render() {
@@ -87,6 +89,15 @@ class Demo extends React.Component {
             <div className='panel-body'>
               <h5>Source in /examples/js/sort/reusable-custom-sort-table.js</h5>
               <ReusableCustomSortTable />
+            </div>
+          </div>
+        </div>
+        <div className='col-md-offset-1 col-md-8'>
+          <div className='panel panel-default'>
+            <div className='panel-heading'>Sort Style Example</div>
+            <div className='panel-body'>
+              <h5>Source in /examples/js/sort/sort-style-table.js</h5>
+              <SortStyleTable />
             </div>
           </div>
         </div>