+++ src/filters/Select.js
@@ -23,9 +23,9 @@ class SelectFilter extends Component {
     };
   }
 
-  componentWillReceiveProps(nextProps) {
-    const isPlaceholderSelected = (nextProps.defaultValue === undefined ||
-      !nextProps.options.hasOwnProperty(nextProps.defaultValue));
+  componentWillReceiveProps() {
+    const currentSelectValue = this.refs.selectInput.value;
+    const isPlaceholderSelected = !currentSelectValue || currentSelectValue === '';
     this.setState(() => {
       return {
         isPlaceholderSelected
@@ -110,7 +110,7 @@ SelectFilter.propTypes = {
   filterHandler: PropTypes.func.isRequired,
   options: PropTypes.object.isRequired,
   placeholder: PropTypes.string,
-  columnName: PropTypes.string,
+  columnName: PropTypes.any,
   style: PropTypes.oneOfType([ PropTypes.object ])
 };