+++ src/util.js
@@ -78,12 +78,15 @@ export default {
     return version === '4';
   },
 
+  isSelectRowDefined(mode) {
+    return mode === Const.ROW_SELECT_SINGLE || mode === Const.ROW_SELECT_MULTI;
+  },
+
   renderColGroup(columns, selectRow, expandColumnOptions = {}, version) {
     let selectRowHeader = null;
     let expandRowHeader = null;
     const isBootstrap4 = this.isBootstrap4(version);
-    const isSelectRowDefined = selectRow.mode === Const.ROW_SELECT_SINGLE ||
-      selectRow.mode === Const.ROW_SELECT_MULTI;
+    const isSelectRowDefined = this.isSelectRowDefined(selectRow.mode);
     const columnWidth = isBootstrap4 ? '38px' : '30px';
     if (isSelectRowDefined) {
       const style = {
@@ -112,7 +115,7 @@ export default {
         not eq offsetWidth in large column table **/
         style.minWidth = width;
       }
-      return (<col style={ style } key={ i } className={ column.className }></col>);
+      return (<col style={ style } key={ i }></col>);
     });
 
     return (